ci: gate the CUDA feature on every PR with a compile-only check (#29)
* ci: gate the CUDA feature on every PR with a compile-only check (sc-11990)

`--features cuda` code is invisible to every other lane: macOS cannot build it,
and the CPU/Linux lanes do not enable the feature. The `windows-cuda` job that
would catch a break is `workflow_dispatch`-only, so nothing checks the feature
unless a human remembers to trigger it.

The consequence is not hypothetical. `candle-gen-mochi`'s `tier_parity.rs`
(sc-11990) merged with three E0308s and was never compiled by anyone — the test
written to validate the tier-ingest seam could not build. It survived a later
commit to the same crate, and was only caught by hand on the CUDA box.

Add `windows-cuda-check`: the same package set and command as the manual lane,
but `--no-run`, and triggered on any PR/push where `select_lanes` marks
`windows_cuda` (candle, contracts, runtime-cuda, workspace-wide changes; MLX-only
and docs PRs never wake the runner).

Two details are load-bearing:

- `--tests` — the break was in a test target, so a bare `cargo build --features
  cuda` would still have missed it.
- `--no-run` — compiling needs nvcc + MSVC but never creates a CUDA context or
  allocates VRAM, so this is safe to run on the shared self-hosted rig without
  competing for the GPU. Executing the CUDA tests stays in the manual lane.

Verified on the self-hosted runner against the real defect:

  broken tier_parity.rs -> GATE_EXIT=101, 3x E0308 (caught)
  fixed  tier_parity.rs -> GATE_EXIT=0, 9 test binaries built (2.47s warm)


* fix(candle-gen-krea): f64 cos-gap band in nvfp4_krea_dit_gpu (sc-11990)

The SC#2 constants MEASURED_COS_GAP/COS_GAP_TOL were f32, but cos_gap is
f64 (cosine() returns f64), so `cos_gap - MEASURED_COS_GAP` failed to
compile (E0308/E0277). This CUDA-only test target was never built by any
lane until the windows-cuda-check gate this PR adds, which caught it on
its first run. Promote the constants to f64 to match the cosine domain;
the {:.5} format specifiers are unaffected. Verified with a CUDA compile.


---------

+++ crates/media/candle-gen/candle-gen-krea/tests/nvfp4_krea_dit_gpu.rs
@@ -700,10 +700,10 @@ fn nvfp4_krea_dit_sc2_parity_vs_q4_tier() {
     // ==========================================================================================
     // Q4 − NVFP4-W4A16 cosine gap measured 2026-07-15 on real weights (this leg quantizes all 260
     // projections; the q4 tier leaves 4 dense — see the surface-mismatch confound in the docs).
-    const MEASURED_COS_GAP: f32 = 0.00601;
+    const MEASURED_COS_GAP: f64 = 0.00601;
     // Absorbs GEMM/library nondeterminism only. The seed, prompt and schedule are fixed, so this is
     // NOT a chaos budget — chaos is *between* the tiers, and it is exactly what the band pins.
-    const COS_GAP_TOL: f32 = 0.004;
+    const COS_GAP_TOL: f64 = 0.004;
 
     let cos_gap = q4_cos - w16_cos;
     assert!(
